requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

# LinkedIn Pipeline Dependencies
anyio==4.9.0
apify_client==1.10.0
apify_shared==1.4.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
dotenv==0.9.9
google-auth==2.40.0
google-auth-oauthlib==1.2.2
gspread==6.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
more-itertools==10.7.0
numpy==2.2.5
oauthlib==3.2.2
pandas==2.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0

# Additional Airflow providers for the pipeline
apache-airflow-providers-postgres
apache-airflow-providers-http
